Move sdk/xlog to sdk/log/xlog
