+++ sdk/log/doc.go
@@ -33,7 +33,7 @@ at a single endpoint their origin is decipherable.
 See [go.opentelemetry.io/otel/log] for more information about
 the OpenTelemetry Logs Bridge API.
 
-See [go.opentelemetry.io/otel/sdk/xlog] for information about the
+See [go.opentelemetry.io/otel/sdk/log/xlog] for information about the
 experimental features.
 */
 package log // import "go.opentelemetry.io/otel/sdk/log"
