+++ sdk/log/go.mod
@@ -10,7 +10,7 @@ require (
 	go.opentelemetry.io/otel v1.34.0
 	go.opentelemetry.io/otel/log v0.10.0
 	go.opentelemetry.io/otel/sdk v1.34.0
-	go.opentelemetry.io/otel/sdk/xlog v0.0.0-00010101000000-000000000000
+	go.opentelemetry.io/otel/sdk/log/xlog v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/otel/trace v1.34.0
 )
 
@@ -30,7 +30,7 @@ replace go.opentelemetry.io/otel/trace => ../../trace
 
 replace go.opentelemetry.io/otel/sdk => ../
 
-replace go.opentelemetry.io/otel/sdk/xlog => ../xlog
+replace go.opentelemetry.io/otel/sdk/log/xlog => ./xlog
 
 replace go.opentelemetry.io/otel/log => ../../log
