+++ sdk/log/processor.go
@@ -13,7 +13,7 @@ import (
 // or with other methods. It is the responsibility of the Processor to manage
 // this concurrency.
 //
-// See [go.opentelemetry.io/otel/sdk/xlog.FilterProcessor] for information about how
+// See [go.opentelemetry.io/otel/sdk/log/xlog.FilterProcessor] for information about how
 // a Processor can support experimental filtering feature.
 type Processor interface {
 	// OnEmit is called when a Record is emitted.
