+++ sdk/log/xlog/doc.go
@@ -7,4 +7,4 @@ These features are added to the OpenTelemetry Go Logs SDK prior to stabilization
 so that users can start experimenting with them and provide feedback.
 These feature may change in backwards incompatible ways as feedback is applied.
 */
-package xlog // import "go.opentelemetry.io/otel/sdk/xlog"
+package xlog // import "go.opentelemetry.io/otel/sdk/log/xlog"
