+++ versions.yaml
@@ -31,7 +31,7 @@ module-sets:
     modules:
       - go.opentelemetry.io/otel/log
       - go.opentelemetry.io/otel/sdk/log
-      - go.opentelemetry.io/otel/sdk/xlog
+      - go.opentelemetry.io/otel/sdk/log/xlog
       - go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc
       - go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp
       - go.opentelemetry.io/otel/exporters/stdout/stdoutlog
